requirements.txt
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
fsspec==2024.9.0
gTTS==2.5.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai-whisper==20231117
pycparser==2.22
pydub==0.25.1
pyOpenSSL==24.2.1
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
six==1.16.0
sympy==1.13.3
tiktoken==0.7.0
torch
tqdm
typing_extensions==4.12.2
urllib3
Werkzeug
